+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/params/query-parameter.ts
@@ -1,5 +1,5 @@
 import type { Parameters } from '../../../../../openapi/types.js'
-import { coerce, stringbool } from '../../../../zod/z/index.js'
+import { coerce, stringbool } from '../../../../zod/utils/index.js'
 
 /**
  * Generates a Zod schema string for a query parameter, with coercion applied based on its type.
