+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/params/request-params-array.ts
@@ -1,4 +1,4 @@
-import { schema } from "../../../../zod/utils/index.js"
+import { schema } from '../../../../zod/utils/index.js'
 
 /**
  * Generates an array of Zod validator strings from OpenAPI parameter objects.
