+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/route-code.ts
@@ -1,5 +1,5 @@
-import { isHttpMethod, isOperation } from '../../../../validator/index.js'
 import type { OpenAPIPaths } from '../../../../openapi/index.js'
+import { isHttpMethod, isOperation } from '../../../../validator/index.js'
 import { route } from './index.js'
 
 /**
