+++ packages/hono-takibi/src/generator/zod/helper/index.ts
@@ -1,5 +1,5 @@
+export { arrayReferenceSchema } from './array-reference-schema.js'
 export { propertiesSchema } from './properties-schema.js'
 export { propertySchema } from './property-schema.js'
-export { arrayReferenceSchema } from './array-reference-schema.js'
 export { referenceSchema } from './reference-schema.js'
 export { zodSchemaFromSubSchema } from './zod-schema-from-sub-schema.js'
