+++ packages/hono-takibi/src/generator/zod/helper/properties-schema.ts
@@ -1,7 +1,7 @@
 import { getToSafeIdentifier } from '../../../core/utils/index.js'
-import { isAllOptional } from '../../../validator/index.js'
 import type { Schema } from '../../../openapi/index.js'
-import { partial } from '../z/index.js'
+import { isAllOptional } from '../../../validator/index.js'
+import { partial } from '../utils/index.js'
 import { propertySchema } from './property-schema.js'
 
 /**
