+++ packages/hono-takibi/src/generator/zod/z/enum.ts
@@ -1,4 +1,4 @@
-import type { Schema } from "../../../openapi/types.js"
+import type { Schema } from '../../../openapi/types.js'
 
 /**
  * Generate a Zod enum schema string from an OpenAPI Schema object.
