+++ packages/hono-takibi/src/generator/zod/z/number.ts
@@ -1,4 +1,3 @@
-
 import type { Schema } from '../../../openapi/types.js'
 import { _default, gt, lt, max, min, regex } from '../utils/index.js'
